fixing warning, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16219 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/emissions/PollutantsInterface.cpp
@@ -56,7 +56,7 @@ PollutantsInterface::Helper* PollutantsInterface::myHelpers[] = {&PollutantsInte
 // ===========================================================================
 SUMOEmissionClass
 PollutantsInterface::getClassByName(const std::string& eClass, const SUMOVehicleClass vc) {
-    int sep = eClass.find("/");
+    size_t sep = eClass.find("/");
     if (sep != std::string::npos) {
         const std::string model = eClass.substr(0, sep);
         const std::string subClass = eClass.substr(sep + 1);
